# docs/swagger.yaml
definitions:
  fail.Response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_request_body.AddCategory:
    properties:
      categoryId:
        type: string
      emoji:
        type: string
      name:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_request_body.UpdateCategory:
    properties:
      emoji:
        type: string
      name:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_response_body.AddCategory:
    properties:
      categoryId:
        type: string
    required:
    - categoryId
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_response_body.Category:
    properties:
      categoryId:
        type: string
      emoji:
        type: string
      name:
        type: string
    required:
    - categoryId
    - name
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_request_body.SetRecipeKey:
    properties:
      key:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_response_body.GetRecipeKey:
    properties:
      key:
        type: string
    type: object
  ? github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_response_body.RecipeKeyRequest
  : properties:
      publicKey:
        type: string
      status:
        type: string
      userAvatar:
        type: string
      userId:
        type: string
      userName:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/encryption/recipes/{recipe_id}:
    get:
      consumes:
      - application/json
      description: Get recipe encrypted key
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_response_body.GetRecipeKey'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get recipe key
      tags:
      - encryption
      - recipe
    post:
      consumes:
      - application/json
      description: Set recipe owner key
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      - description: Key
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_request_body.SetRecipeKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Set recipe owner key
      tags:
      - encryption
      - recipe
  /v1/encryption/recipes/{recipe_id}/users:
    get:
      consumes:
      - application/json
      description: Get recipe key access requests
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_response_body.RecipeKeyRequest'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get recipe key requests
      tags:
      - encryption
      - recipe
    post:
      consumes:
      - application/json
      description: Request recipe key access
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Request recipe key access
      tags:
      - encryption
      - recipe
  /v1/encryption/recipes/{recipe_id}/users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete recipe key
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete recipe key
      tags:
      - encryption
      - recipe
    post:
      consumes:
      - application/json
      description: Set recipe user key
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Key
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_encryption_dto_request_body.SetRecipeKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Set recipe user key
      tags:
      - encryption
      - recipe
...
  /v1/recipes/categories:
    get:
      consumes:
      - application/json
      description: Get user recipe categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_response_body.Category'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user categories
      tags:
      - category
      - recipe
    post:
      consumes:
      - application/json
      description: Add recipe category
      parameters:
      - description: Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_request_body.AddCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_response_body.AddCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Add category
      tags:
      - category
      - recipe
  /v1/recipes/categories/{category_id}:
    delete:
      consumes:
      - application/json
      description: Delete recipe category
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete category
      tags:
      - category
    get:
      consumes:
      - application/json
      description: Get recipe category
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_response_body.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get category
      tags:
      - category
      - recipe
    put:
      consumes:
      - application/json
      description: Update recipe category
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: integer
      - description: Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_category_dto_request_body.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Update category
      tags:
      - category
      - recipe
...
swagger: "2.0"